fix(date_picker): REFACTOR static methods to diff classes
