+++ DashMonitor/app/views/layouts/benchmark_analysis.py
@@ -15,6 +15,7 @@ from dash import html
 
 # local imports
 from DashMonitor.app.data.analyzers import TimeTrendAnalyzer
+from DashMonitor.app.handlers.date_utils import DateUtils
 from DashMonitor.app.handlers.function_utils import (
     categorize_score,
     categorize_score_to_text_class_name,
@@ -27,6 +28,7 @@ from DashMonitor.app.views.components.date_picker.callbacks import (
     register_datepicker_callbacks,
 )
 from DashMonitor.app.views.components.date_picker.date_picker import DatePicker
+from DashMonitor.app.views.components.utils.button import ButtonUtils
 from DashMonitor.app.views.layouts.mock_data_sasb_analysis import *
 from DashMonitor.app.views.configs import (
     main_df_provider,
@@ -293,9 +295,9 @@ def register_callbacks(app):
             end_date = pd.to_datetime(end_date)
         else:
             # If a predefined button is selected, determine the corresponding date range
-            selected_index = DatePicker.selected_button_index(button_classes)
+            selected_index = ButtonUtils.selected_button_index(button_classes)
             end_date = pd.to_datetime(date.today())
-            start_date = pd.to_datetime(DatePicker.calculate_start_date(selected_index))
+            start_date = pd.to_datetime(DateUtils().calculate_start_date_on_index(selected_index))
 
         # Aggregate sentiment scores by bank name and date
         overall_df = (
